Time: 1183 ms (7.68%), Space: 423.1 MB (8.2%) - LeetHub

+++ 2170-count-number-of-maximum-bitwise-or-subsets/2170-count-number-of-maximum-bitwise-or-subsets.cpp
@@ -32,7 +32,7 @@
 
 class Solution {
 public:
-    int solve (int idx,int curOr,vector<int> nums,short sizeOfArray,int maxOr)
+    int solve (short idx,int curOr,vector<int> nums,short sizeOfArray,int maxOr)
     {
         if(idx==sizeOfArray)
         return maxOr==curOr;
